Fix rebase when there are symlinks in .git/ on Windows
When performing a rebase, rmdir() is called on the folder .git/logs
which is supposed to exit without deleting anything in case .git/logs is
a symbolic link. The Windows equivalent to rmdir() (_rmdir) does not
behave the same and removes the folder if it is symlink which generates
issues especially when Repo is used.
This commit updates mingw_rmdir() so that its behavior is the same as
rmdir() in case of symbolic links.

Signed-off-by: Thomas Bétous <tomspycell@gmail.com>